bonding implemented, added sigmoid reward scaling
