+++ storage/validator/__init__.py
@@ -22,3 +22,5 @@ from . import utils
 from . import verify
 from . import encryption
 from . import database
+from . import reward
+from . import bonding
